Add 16x16 grid and red etch-a-sketch border
This commit uses index, styles and javascript to make and format a 16x16
grid. In the index files, it adds a container div which forms the outer
shell of the etch-a-sketch. It also has 16 row divs.
Those 16 row divs are then used in javscript with loops to provide
housing for the 16 divs going horizontally. This completes the grid.
In styles.css, the borders and sizing is set for these different divs.
styles.css also sets the row divs as a row-direction flexbox to ensure
the divs added in javascript are placed horizontally rather than
vertically.