Add slider to customize grid size
This commit adds a slider to the ui that allows users to customize the
size of the grid.
It will default to a 32x32 grid and allow users to slide for any size
between 16 x 16 and 100 x 100.
With this change, I removed the row divs from .container. This will
instead be handled dynamically with javascript.