Add function to remove grid before adding new grid
The custom grid function based on slider was only functional the first
time the slider was used. Additional uses of the slider would add the
new grid to the old grid and disrupt the layout of the page.
This commit adds a function that removes the grid to replace it with a
new one when the slider is used.