Adjust grid slider functionality to be more fluid
This commit changes the grid function event listener from 'click' to
'mousemove'. When 'click' was being used, users could only see the grid
being selected after the selection was made.
Now when the slider is used, the grid and the display both show users
exactly what the grid looks like as they are sliding through the sizes.
This is much more fluid and will help users select a grid size without
guesswork.