Break apart the gridSlider event listeners
The 'mousemove' event listener on the gridSlider was giving the ability
for users to make a more informed decision on their grid size, but it
negative impact in the fact that it was also causing the grid to clear
if the mouse accidentally crossed over it.
To resolve this, I broke the event listener down to 2 event listeners.
One of these event listeners checks form mouse movement and adjusts the
display where the grid size is typed out for users. This effectively
allows users to see the sizes as they use the slider. However, the grid
itself will only adjust when a click is made on the slider.