Add shading tool functionality
This commit adds functionality to the shading tool. Each div within the
etch-a-sketch frame now has a natural opacity of 10%. When the color or
color at random tools are used, the opacity is set to 100%. when the
shader tool is used, the opacity goes up 10% on each pass.
With this, the borders were changing opacity. I resolved this by
removing the borders.