Change background color of etch-a-sketch to white
When I made the etch-a-sketch divs have 10% opacity, the background of
the etch-a-sketch became the same as the site background, an off-white.
This made use of the eraser make the div brighter than untouched divs.
I adjusted the background of the frame that holds the interactive divs
to be an opaque white, which made the divs over it white without
adjusting their opacity.