Adjust clearing of the divs in js
Found that after you clear the etch-a-sketch, you coluld not draw on it
again. The problem was a lack of reestablishment on the divs declaration,
which caused it to not have the event listener applied for when the
mouse moves over the divs. This would need to be reestablished because
I was deleting the divs and making new ones. The same thing happens when
the grid is changed, which determined a need to have the event listener
for mouse moves over the divs nested within the event listener for the\
grid slider being adjusted.
Rather than nesting this event listener again, I changed how the board
resets, instead just setting up a loop that sets all divs to white and
10% opacity.